requirements.txt
uvicorn
fastapi
pydantic
langchain-huggingface
langchain-chroma
langchain-text-splitters
sentence_transformers
praw
python-dotenv
supabase
pydantic[email]
openai
python-jose[cryptography]
python-multipart
bcrypt